Fixed a syntax error in the JSON that would cause version check errors.
Both the 1.20.2-recommended and 1.20.4-recommended version entries lacked a trailing comma, likely as a result of a copy-paste error when the 1.20 versions were added to the list. This would cause the received JSON to be malformed, which would prevent a version check from taking place in either the client or the server.